+++ packages/xpce/src/ker/alloc.h
@@ -46,7 +46,7 @@ struct zone
 
 GLOBAL Zone freeChains[ALLOCFAST/sizeof(Zone)+1];
 
-#define struct_offset(structure, field) ((int) &(((structure *)NULL)->field))
+#define struct_offset(structure, field) ((size_t) &(((structure *)NULL)->field))
 #define MINALLOC    (sizeof(struct zone) - struct_offset(struct zone, start))
 #define ROUNDALLOC  (sizeof(void *))
